requirements.txt
annotated-types==0.7.0
anyio==4.4.0
boto3==1.34.157
botocore==1.34.157
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.112.0
h11==0.14.0
idna==3.7
jmespath==1.0.1
mangum==0.17.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
